// frontend/src/api/schemas/config.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Prism Platform API
 * API for managing traces, configurations, devices, and queries
 * OpenAPI spec version: 1.0.0
 */

export interface Config {
  config_id?: string;
  config_name?: string;
  config_text?: string;
  tracing_tool?: string;
  default_duration?: number;
  updated_at?: string;
}