// frontend/src/api/schemas/query.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Prism Platform API
 * API for managing traces, configurations, devices, and queries
 * OpenAPI spec version: 1.0.0
 */

export interface Query {
  query_id?: string;
  query_name?: string;
  query_text?: string;
  configuration_id?: string;
  updated_at?: string;
}